function to allocate memory for an array of nmemb elements of size bytes each and returns a pointer to the allocated memory.

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -19,7 +19,7 @@ void *_calloc(unsigned int nmemb, unsigned int size)
 		return (NULL);
 	if (size == 0 || nmemb == 0)
 		return (NULL);
-	for (i = 0; i < nmemb; i++)
+	for (i = 0; i < (int)nmemb; i++)
 		ar[i] = 0;
 	return (ar);
 }
